requirements.txt
python==3.8.20
pytorch==2.3.1
CUDA==11.8
numpy==1.21.5
scikit-learn==0.22.2.post1
scipy==1.7.3
